+++ R/gg_point.R
@@ -455,7 +455,7 @@ gg_point_col <- function(data,
                                                                  ordered_result = TRUE,
                                                                  format_fun = col_labels)))
     
-    col_labels <- sv_interval_breaks_to_interval_labels
+    col_labels <- sv_interval_labels_chr
     
     col_n <- length(col_cuts) - 1
     if (is.null(pal)) pal <- pal_viridis_reorder(col_n)
@@ -1121,7 +1121,7 @@ gg_point_col_facet <-
                                                                    ordered_result = TRUE,
                                                                    format_fun = col_labels)))
       
-      col_labels <- sv_interval_breaks_to_interval_labels
+      col_labels <- sv_interval_labels_chr
       
       col_n <- length(col_cuts) - 1
       if (is.null(pal)) pal <- pal_viridis_reorder(col_n)